refactor: rename 'org' to 'organizationName' and implement filtering and searching functionality in admin users page
